% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_xlsx.R
\name{add_TOC_sheet}
\alias{add_TOC_sheet}
\title{Add table of contents to an workbook}
\usage{
add_TOC_sheet(wb, link_text = NULL, return_pos = c(1, 1))
}
\arguments{
\item{wb}{the workbook to append the sheet to (requires append_to = TRUE)}

\item{link_text}{character vector of the text displayed for the linked text in the TOC sheet. Must have the same length as the number of sheets in wb.}

\item{return_pos}{integer pair, of the row and column index to place the return to contents link on the sheets in the wb.}
}
\description{
Add table of contents to an workbook
}
