% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_xlsx.R
\name{to_sheet_theme_summary}
\alias{to_sheet_theme_summary}
\title{Add theme summary sheet}
\usage{
to_sheet_theme_summary(
  data,
  theme = NULL,
  theme_columns = NULL,
  append_to = NULL,
  wb = NULL,
  sheet = "Data",
  order = FALSE
)
}
\arguments{
\item{data}{survey data, formatted where each row is an individual and column names describe demographics or survey questions (of the format YY__ZZ, where XX is the theme and YY the question, where ' ' are replaced with '_' and '?' with 'XX'.)}

\item{theme}{string containing the theme name.}

\item{theme_columns}{Indices of the columns in the data for the theme}

\item{append_to}{Flag (TRUE/FALSE) for whether the summary sheet is appended to a previous workbook}

\item{wb}{the workbook to append the sheet to (requires append_to = TRUE)}

\item{sheet}{the sheet name}

\item{order}{not used}
}
\value{
a workbook with the theme summary added
}
\description{
Add theme summary sheet
}
