% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{year7_pilot}
\alias{year7_pilot}
\title{Fake survey data}
\format{
An object of class \code{data.frame} with 500 rows and 65 columns.
}
\usage{
year7_pilot
}
\description{
Randomly generated data for 500 individuals if they were to partake in the year 7 pilot survey.
}
\details{
#' A data frame with 500 rows and 65 columns:
}
\keyword{datasets}
